refactor: move DriverList component to pages directory and update import path
