+++ src/pages/DriverList.tsx
@@ -1,4 +1,4 @@
-import Card from './Card'
+import Card from "../components/Card"
 
 const DriverList = () => {
   return (
